requirements.txt
selenium
pytest-bdd
webdriver-manager
requests
pytest-html-reporter
